+++ RELEASENOTES.md
@@ -1,3 +1,9 @@
+# v1.6.0 - TBD
+
+- dataKeyPieceOverride functionality is removed
+- paths like 'a.b.c' where 'b' is array is not supported. use a.b[*].c with jsonave.
+- 'dataKeyFnOptions' replaced with 'context' option.
+
 # v1.5.0 - June 12, 2015
 
 - `dataKey` now accepts functions.  Primarily designed for JSONPath expression.